Moved platform to attic
