--- java/spring/spring-webapp/src/main/java/net/ivoa/calycopis/functional/platfom/canfar/CanfarExecutionSession.java
@@ -1,15 +1,0 @@
-/**
- * 
- */
-package net.ivoa.calycopis.functional.platfom.canfar;
-
-import net.ivoa.calycopis.functional.platfom.ConfigStep;
-
-/**
- * 
- */
-public interface CanfarExecutionSession
-    extends ConfigStep
-    {
-
-    }
