+++ java/spring/spring-webapp/src/main/java/net/ivoa/calycopis/functional/platfomattic/ConfigStepEntity.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package net.ivoa.calycopis.functional.platfom;
+package net.ivoa.calycopis.functional.platfomattic;
 
 import java.time.Duration;
