+++ java/spring/spring-webapp/src/main/java/net/ivoa/calycopis/functional/platfomattic/canfar/CanfarConfigStep.java
@@ -1,9 +1,9 @@
 /**
  * 
  */
-package net.ivoa.calycopis.functional.platfom.canfar;
+package net.ivoa.calycopis.functional.platfomattic.canfar;
 
-import net.ivoa.calycopis.functional.platfom.ConfigStep;
+import net.ivoa.calycopis.functional.platfomattic.ConfigStep;
 
 /**
  *
