+++ java/spring/spring-webapp/src/main/java/net/ivoa/calycopis/functional/platfomattic/canfar/CanfarConfigStepFactory.java
@@ -1,11 +1,11 @@
 /**
  * 
  */
-package net.ivoa.calycopis.functional.platfom.canfar;
+package net.ivoa.calycopis.functional.platfomattic.canfar;
 
 import net.ivoa.calycopis.datamodel.session.ExecutionSessionEntity;
-import net.ivoa.calycopis.functional.platfom.ConfigStepEntity;
-import net.ivoa.calycopis.functional.platfom.ConfigStepFactory;
+import net.ivoa.calycopis.functional.platfomattic.ConfigStepEntity;
+import net.ivoa.calycopis.functional.platfomattic.ConfigStepFactory;
 
 /**
  *
