+++ java/spring/spring-webapp/src/main/java/net/ivoa/calycopis/functional/platfomattic/canfar/CanfarExecutionSession.java
@@ -1,0 +1,15 @@
+/**
+ * 
+ */
+package net.ivoa.calycopis.functional.platfomattic.canfar;
+
+import net.ivoa.calycopis.functional.platfomattic.ConfigStep;
+
+/**
+ * 
+ */
+public interface CanfarExecutionSession
+    extends ConfigStep
+    {
+
+    }
