+++ java/spring/spring-webapp/src/main/java/net/ivoa/calycopis/functional/platfomattic/canfar/CanfarExecutionSessionEntity.java
@@ -1,16 +1,16 @@
 /**
  * 
  */
-package net.ivoa.calycopis.functional.platfom.canfar;
+package net.ivoa.calycopis.functional.platfomattic.canfar;
 
 import java.time.Duration;
 
 import lombok.extern.slf4j.Slf4j;
 import net.ivoa.calycopis.datamodel.session.ExecutionSessionEntity;
-import net.ivoa.calycopis.functional.platfom.ConfigStepEntity;
-import net.ivoa.calycopis.functional.platfom.vospace.VOSpaceConfigStep;
-import net.ivoa.calycopis.functional.platfom.vospace.VOSpaceFileImportEntity;
-import net.ivoa.calycopis.functional.platfom.vospace.VOSpaceUserDirectoryEntity;
+import net.ivoa.calycopis.functional.platfomattic.ConfigStepEntity;
+import net.ivoa.calycopis.functional.platfomattic.vospace.VOSpaceConfigStep;
+import net.ivoa.calycopis.functional.platfomattic.vospace.VOSpaceFileImportEntity;
+import net.ivoa.calycopis.functional.platfomattic.vospace.VOSpaceUserDirectoryEntity;
 
 /**
  *
