+++ java/spring/spring-webapp/src/main/java/net/ivoa/calycopis/functional/platfomattic/vospace/VOSpaceFileImportConfig.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package net.ivoa.calycopis.functional.platfom.vospace;
+package net.ivoa.calycopis.functional.platfomattic.vospace;
 
 /**
  *
