+++ java/spring/spring-webapp/src/main/java/net/ivoa/calycopis/functional/platfomattic/vospace/VOSpaceUserDirectoryEntity.java
@@ -1,12 +1,12 @@
 /**
  * 
  */
-package net.ivoa.calycopis.functional.platfom.vospace;
+package net.ivoa.calycopis.functional.platfomattic.vospace;
 
 import java.time.Duration;
 
 import lombok.extern.slf4j.Slf4j;
-import net.ivoa.calycopis.functional.platfom.ConfigStepEntity;
+import net.ivoa.calycopis.functional.platfomattic.ConfigStepEntity;
 
 /**
  * Public interface for a StorageVolume mount in CANFAR.
